USA/MINNESOTA/requirements.txt
requests==2.31.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
urllib3==1.26
boto3==1.34.63
beautifulsoup4==4.12.3
lxml==5.2.1